gen-ai/image_upscale/model_code/requirements.txt
# Core diffusion & Hugging Face
diffusers
transformers
accelerate
safetensors

# Image processing
pillow